OpenAIBasicChatGPT/requirements.txt
openai
langchain_openai
streamlit
python-dotenv